+++ crates/orchestrator/src/tests/common/mod.rs
@@ -1,22 +1,23 @@
 pub mod constants;
 
 use std::collections::HashMap;
+use std::sync::Arc;
 
+use crate::config::{get_aws_config, ProviderConfig};
 use ::uuid::Uuid;
 use aws_config::Region;
 use aws_sdk_sns::error::SdkError;
 use aws_sdk_sns::operation::create_topic::CreateTopicError;
+use chrono::{SubsecRound, Utc};
 use mongodb::Client;
 use rstest::*;
 use serde::Deserialize;
 use utils::env_utils::get_env_var_or_panic;
+use utils::settings::env::EnvSettingsProvider;
 
-use crate::data_storage::aws_s3::config::AWSS3Config;
 use crate::data_storage::aws_s3::AWSS3;
-use crate::data_storage::{DataStorage, DataStorageConfig};
-use crate::database::mongodb::config::MongoDbConfig;
+use crate::data_storage::DataStorage;
 use crate::database::mongodb::MongoDb;
-use crate::database::DatabaseConfig;
 use crate::jobs::types::JobStatus::Created;
 use crate::jobs::types::JobType::DataSubmission;
 use crate::jobs::types::{ExternalId, JobItem};
@@ -32,6 +33,8 @@ pub fn default_job_item() -> JobItem {
         external_id: ExternalId::String("0".to_string().into_boxed_str()),
         metadata: HashMap::new(),
         version: 0,
+        created_at: Utc::now().round_subsecs(0),
+        updated_at: Utc::now().round_subsecs(0),
     }
 }
 
@@ -56,7 +59,7 @@ pub async fn get_sns_client() -> aws_sdk_sns::client::Client {
 }
 
 pub async fn drop_database() -> color_eyre::Result<()> {
-    let db_client: Client = MongoDb::new(MongoDbConfig::new_from_env()).await.client();
+    let db_client: Client = MongoDb::new_with_settings(&EnvSettingsProvider {}).await.client();
     // dropping all the collection.
     // use .collection::<JobItem>("<collection_name>")
     // if only particular collection is to be dropped
@@ -99,7 +102,11 @@ pub struct MessagePayloadType {
 }
 
 pub async fn get_storage_client() -> Box<dyn DataStorage + Send + Sync> {
-    let aws_config =
-        aws_config::load_from_env().await.into_builder().endpoint_url(get_env_var_or_panic("AWS_ENDPOINT_URL")).build();
-    Box::new(AWSS3::new(AWSS3Config::new_from_env(), &aws_config))
+    Box::new(
+        AWSS3::new_with_settings(
+            &EnvSettingsProvider {},
+            ProviderConfig::AWS(Arc::new(get_aws_config(&EnvSettingsProvider {}).await)),
+        )
+        .await,
+    )
 }